requirements.txt
astroid==2.5.2
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==6.7
Flask==1.1.2
idna==2.7
isort==5.8.0
itsdangerous==0.24
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==0.23
mccabe==0.6.1
plaid-python==5.0.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.7.3
PyNaCl==1.4.0
python-dotenv==0.14.0
requests==2.20.0
six==1.15.0
telnyx==1.2.2
toml==0.10.2
urllib3==1.24.3
Werkzeug==1.0.1
wrapt==1.12.1
